+++ src/widgets/filenavigationwidget.h
@@ -25,21 +25,6 @@ class FileNavigationWidget : public QTreeWidget {
     Q_OBJECT
 
    public:
-    explicit FileNavigationWidget(QWidget *parent = 0);
-
-    void parse(const QTextDocument *document, int textCursorPosition);
-    void selectItemForCursorPosition(int position);
-    bool hasFileLinks() const;
-
-   private slots:
-    void onCurrentItemChanged(QTreeWidgetItem *current, QTreeWidgetItem *previous);
-    void onItemClicked(QTreeWidgetItem *current, int column);
-    void showContextMenu(const QPoint &pos);
-
-   signals:
-    void positionClicked(int position);
-
-   private:
     enum class FileItemType { MediaFile, AttachmentFile };
 
     struct FileLinkNode {
@@ -54,11 +39,27 @@ class FileNavigationWidget : public QTreeWidget {
         }
     };
 
+    explicit FileNavigationWidget(QWidget *parent = 0);
+
+    void parse(const QTextDocument *document, int textCursorPosition);
+    void setFileLinkNodes(const QVector<FileLinkNode> &nodes, int textCursorPosition);
+    void selectItemForCursorPosition(int position);
+    bool hasFileLinks() const;
+    static QVector<FileLinkNode> parseText(const QString &text, const QString &mediaPath,
+                                           const QString &attachmentsPath);
+
+   private slots:
+    void onCurrentItemChanged(QTreeWidgetItem *current, QTreeWidgetItem *previous);
+    void onItemClicked(QTreeWidgetItem *current, int column);
+    void showContextMenu(const QPoint &pos);
+
+   signals:
+    void positionClicked(int position);
+
+   private:
     void buildTree(const QVector<FileLinkNode> &nodes);
-    static QVector<FileLinkNode> parseDocument(const QTextDocument *document);
     void emitPositionForItem(const QTreeWidgetItem *item);
 
-    const QTextDocument *_doc = nullptr;
     QVector<FileLinkNode> _fileLinkNodes;
     int _cursorPosition = 0;
 };